UserStory #15 Completed. Added Tooltip for city views
