+++ src/pages/Home.js
@@ -1,6 +1,6 @@
 import "../style/Home.css"
 import SearchBarComp from "../components/searchBar";
-import CityInfoHeader from "../components/CityInfoHeader";
+import CityInfoHeader from "../components/cityInfoHeader";
 import {useState} from "react";
 import axios from "axios";
 
@@ -9,7 +9,7 @@ function Home() {
     const [cityData, setCityData] = useState([])
 
     function getAllImages() {
-        let APICallString = "http://localhost:8080/api/v1/images/getByTitle/Thessaloniki";
+        let APICallString = "http://localhost:8080/api/v1/images/getByTitle/" + cityProp;
         axios.get(APICallString).then(function (response) {
             const data = response.data
             setCityData(data)
