+++ src/pages/city/[id]/singlePage.js
@@ -1,0 +1,13 @@
+
+
+function SinglePage(){
+    return(
+        <>
+            <div>
+                <h1>TEST SINGLE PAGE</h1>
+            </div>
+        </>
+    )
+}
+
+export default SinglePage
